requirements.txt
feedparser==5.2.1
python-dateutil==2.8.0
six==1.12.0
